refactor(data): move default binarizer to its own file
